Reorganizing files to be multi-light friendly
