+++ TODO.md
@@ -1,0 +1,4 @@
+* Work with the new node-hue-api library
+* Abstract lights from being Hue-specific
+* Teach system to talk to FastLED
+* Fewer callbacks, more async/await
