+++ lib/light-watcher.js
