+++ lib/lights/index.js
@@ -1,0 +1,2 @@
+exports.hue = require('./hue/')
+exports.fastled = require('./fastled/')
